// types.ts
export enum Tab {
  CHATBOT = 'Chatbot',
  GENERAL_TASKS = 'General Tasks',
  IMAGE_ANALYZER = 'Image Analyzer',
  IMAGE_EDITOR = 'Image Editor',
  IMAGE_GENERATOR = 'Image Generator',
  LIVE_CONVERSATION = 'Live Conversation',
  GROUNDED_SEARCH = 'Grounded Search',
  TEXT_TO_SPEECH = 'Text-to-Speech',
  VIDEO_ANALYZER = 'Video Analyzer',
  VIDEO_GENERATOR = 'Video Generator',
}